+++ src-tauri/entity/src/relacion_venta_prod.rs
@@ -8,7 +8,7 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub producto: i32,
-    pub cantidad: i32,
+    pub cantidad: u8,
     #[sea_orm(column_type = "Double")]
     pub precio: f32,
     pub venta: i32,
