+++ src-tauri/src/mods/relacion_prod_prov.rs
@@ -12,17 +12,17 @@ pub struct RelacionProdProv {
 }
 
 impl RelacionProdProv {
-    pub fn new(id_producto: i64, id_proveedor: i64, codigo_interno: Option<i64>) -> Self {
+    pub fn new(id_producto: i32, id_proveedor: i32, codigo_interno: Option<i64>) -> Self {
         RelacionProdProv {
             id_producto,
             id_proveedor,
             codigo_interno,
         }
     }
-    pub fn id_producto(&self) -> &i64 {
+    pub fn id_producto(&self) -> &i32 {
         &self.id_producto
     }
-    pub fn id_proveedor(&self) -> &i64 {
+    pub fn id_proveedor(&self) -> &i32 {
         &self.id_proveedor
     }
     pub fn codigo_interno(&self) -> Option<i64> {
